# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kizchat social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string }
                bio: { type: string }
                profilePicture: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserSummary" }
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
  /users/{username}/block:
    post:
      operationId: blockUser
      tags: [privacy]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockState" }
    delete:
      operationId: unblockUser
      tags: [privacy]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockState" }
  /me/privacy:
    patch:
      operationId: updatePrivacy
      tags: [privacy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lastSeenVisible: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
  /users/{username}/follow:
    post:
      operationId: followUser
      tags: [users]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FollowState" }
    delete:
      operationId: unfollowUser
      tags: [users]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FollowState" }
  /users/{username}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserSummary" }
  /users/{username}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserSummary" }
  /users/{username}/videos:
    get:
      operationId: getUserVideos
      tags: [videos]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Video" }
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id: { type: integer }
        clerkId: { type: string }
        username: { type: string }
        displayName: { type: string }
        bio: { type: string }
        profilePicture: { type: string }
        isVerified: { type: boolean }
        verificationRequested: { type: boolean }
        followersCount: { type: integer }
        followingCount: { type: integer }
        videosCount: { type: integer }
        lastSeenVisible: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, username, displayName, bio, profilePicture, isVerified, verificationRequested, followersCount, followingCount, videosCount, lastSeenVisible, createdAt]
    UserSummary:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        bio: { type: string }
        profilePicture: { type: string }
        isVerified: { type: boolean }
        followersCount: { type: integer }
        isFollowing: { type: boolean }
      required: [id, username, displayName, bio, profilePicture, isVerified, followersCount, isFollowing]
    UserProfile:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        bio: { type: string }
        profilePicture: { type: string }
        isVerified: { type: boolean }
        followersCount: { type: integer }
        followingCount: { type: integer }
        videosCount: { type: integer }
        likesCount: { type: integer }
        isFollowing: { type: boolean }
        isMe: { type: boolean }
      required: [id, username, displayName, bio, profilePicture, isVerified, followersCount, followingCount, videosCount, likesCount, isFollowing, isMe]
    FollowState:
      type: object
      properties:
        isFollowing: { type: boolean }
        followersCount: { type: integer }
      required: [isFollowing, followersCount]
    Video:
      type: object
      properties:
        id: { type: integer }
        videoUrl: { type: string }
        thumbnailUrl: { type: string }
        caption: { type: string }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        liked: { type: boolean }
        createdAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/UserSummary" }
      required: [id, videoUrl, thumbnailUrl, caption, likesCount, commentsCount, liked, createdAt, author]
...
    BlockState:
      type: object
      properties:
        isBlocked: { type: boolean }
      required: [isBlocked]
...