# lib/api-zod/src/generated/types/followState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kizchat social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FollowState {
  isFollowing: boolean;
  followersCount: number;
}