# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kizchat social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  clerkId?: string;
  username: string;
  displayName: string;
  bio: string;
  profilePicture: string;
  isVerified: boolean;
  verificationRequested: boolean;
  followersCount: number;
  followingCount: number;
  videosCount: number;
  lastSeenVisible: boolean;
  createdAt: Date;
}